refactor(watch): extract useChannelAutopilot; useChannels is now a thin orchestrator
